requirements.txt
streamlit
rembg
pillow